requirements.txt
boto3==1.5.12
botocore==1.8.26
certifi==2017.11.5
chardet==3.0.4
cssselect==1.0.1
dj-database-url==0.4.2
Django==1.11.7
django-storages==1.6.5
docopt==0.6.2
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
housecanary==0.7.1
idna==2.6
jdcal==1.3
jmespath==0.9.3
lxml==4.1.1
openpyxl==2.4.9
probableparsing==0.0.1
psycopg2==2.7.3.2
pyquery==1.3.0
python-crfsuite==0.9.5
python-dateutil==2.6.1
python-google-places==1.4.1
python-slugify==1.2.4
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
usaddress==0.5.10
waitress==1.1.0
